Rename defaults packages to specific defaults
