+++ cmd/goa4web/dbinit.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"github.com/arran4/goa4web/internal/dbdrivers/allstable"
+	dbdefaults "github.com/arran4/goa4web/internal/dbdrivers/dbdefaults"
 )
 
 func init() {
-	allstable.Register()
+	dbdefaults.Register()
 }
