+++ cmd/goa4web/main.go
@@ -9,11 +9,11 @@ import (
 	"log"
 	"os"
 
-	_ "github.com/arran4/goa4web/cmd/goa4web/dbhandlers/allstable"
+	_ "github.com/arran4/goa4web/cmd/goa4web/dbhandlers/dbdefaults"
 	"github.com/arran4/goa4web/config"
 	"github.com/arran4/goa4web/core"
 	dbstart "github.com/arran4/goa4web/internal/dbstart"
-	dlqreg "github.com/arran4/goa4web/internal/dlq/register/defaults"
+	dlqreg "github.com/arran4/goa4web/internal/dlq/dlqdefaults"
 	"github.com/arran4/goa4web/runtimeconfig"
 )
