+++ internal/dlq/dlq_test.go
@@ -8,14 +8,14 @@ import (
 	dlq "github.com/arran4/goa4web/internal/dlq"
 	dbdlq "github.com/arran4/goa4web/internal/dlq/db"
 	dirdlq "github.com/arran4/goa4web/internal/dlq/dir"
+	dlqdefaults "github.com/arran4/goa4web/internal/dlq/dlqdefaults"
 	emaildlq "github.com/arran4/goa4web/internal/dlq/email"
 	filedlq "github.com/arran4/goa4web/internal/dlq/file"
-	defaults "github.com/arran4/goa4web/internal/dlq/register/defaults"
 	"github.com/arran4/goa4web/runtimeconfig"
 )
 
 func TestProviderFromConfigRegistry(t *testing.T) {
-	defaults.Register()
+	dlqdefaults.Register()
 
 	cfg := runtimeconfig.RuntimeConfig{DLQProvider: "file", DLQFile: "p"}
 	if _, ok := dlq.ProviderFromConfig(cfg, nil).(*filedlq.DLQ); !ok {
