+++ internal/dlq/dlqdefaults/defaults.go
@@ -1,4 +1,4 @@
-package defaults
+package dlqdefaults
 
 import (
 	"github.com/arran4/goa4web/internal/dlq/db"
