+++ internal/email/emaildefaults/builtin.go
@@ -1,4 +1,4 @@
-package builtin
+package emaildefaults
 
 import (
 	"github.com/arran4/goa4web/internal/email/jmap"
